refactor(session_resources): absorb SessionResourceService into queries
The service had one 40-line method `list_for_session` called from exactly
one command and nowhere else. Inline it into `queries::list_for_session`
and call it directly from `ListSessionResources::execute`. Drops
`SessionResourceService` from `Ctx` and `api::session_resources::AppState`
(neither was read after construction).

Net -70 LOC. No behavioral change.

Picks the smallest internal `service.rs` first as a template; the larger
orchestration services (sessions, mcp_servers, evals, budgets, etc.)
remain under the three-tier shape documented in `specs/domains.md`.
