+++ crates/server/src/domains/session_resources/mod.rs
@@ -1,13 +1,7 @@
-// Session resources domain — commands, queries, types.
-//
-// See specs/domains.md for the pattern.
-
 pub mod commands;
 pub mod leased_resource;
 pub mod queries;
-pub mod service;
 pub mod types;
 
 pub use commands::*;
 pub use leased_resource::LeasedResourceService;
-pub use service::*;
